+++ pywren_ibm_cloud/compute/backends/ibm_cf/ibm_cf.py
@@ -226,8 +226,10 @@ class IBMCloudFunctionsBackend:
         """
         action_name = self._format_action_name(docker_image_name, runtime_memory)
 
-        activation_id = self.cf_client.invoke(self.package, action_name,
-                                              payload, self.is_pywren_function)
+        activation_id = self.cf_client.invoke(package=self.package,
+                                              action_name=action_name,
+                                              payload=payload,
+                                              is_ow_action=self.is_pywren_function)
 
         return activation_id