+++ pywren_ibm_cloud/compute/backends/openwhisk/config.py
@@ -6,7 +6,7 @@ RUNTIME_DEFAULT_35 = 'ibmfunctions/pywren:3.5'
 RUNTIME_DEFAULT_36 = 'ibmfunctions/action-python-v3.6'
 RUNTIME_DEFAULT_37 = 'ibmfunctions/action-python-v3.7:1.6.0'
 
-RUNTIME_TIMEOUT_DEFAULT = 300000  # Default: 600000 milliseconds => 10 minutes
+RUNTIME_TIMEOUT_DEFAULT = 300000  # Default: 300000 milliseconds => 5 minutes
 RUNTIME_MEMORY_DEFAULT = 256  # Default memory: 256 MB
 MAX_CONCURRENT_WORKERS = 100